Get a list of propositions discussed in between a time interval

Get a list of propositions discussed in between a time interval

proposicoes(..., from, to)

Arguments

...
query parameters for the House of Representatives API (See: https://dadosabertos.camara.leg.br/swagger/api.html)

from
start date (YYYY-MM-DD) for the time interval in which a proposition was discussed

to
end date (YYYY-MM) for the time interval in which a proposition was discussed

Value

A tibble of propositions discussed in between a time interval

See also

Examples

proposicoes(from = "2020-01-01", to = "2020-12-01", itens = 100)

#> # A tibble: 100 x 6
#>       id siglaTipo codTipo numero   ano ementa                                  
#>    <int> <chr>       <int>  <int> <int> <chr>                                   
#>  1 15009 PL            139     70  1995 "Dispõe sobre intervenções cirúrgicas q~
#>  2 15532 PL            139    505  1991 "Revoga a alínea \"l\" do art. 20 do De~
#>  3 15749 PL            139    693  1999 "Altera a Lei nº 8.245, de 18 de outubr~
#>  4 15990 PL            139    887  1991 "Cria salvaguardas para a tecnologia no~
#>  5 16481 PL            139   1258  1995 "Disciplina o inciso XII do art. 5º da ~
#>  6 16969 PL            139   1610  1996 "Dispõe sobre a exploração e o aproveit~
#>  7 17563 PL            139   2051  1996 "Dispõe sobre a assistência médica, hos~
#>  8 17823 PL            139   2231  1999 "Obriga os responsáveis por \"sites\" p~
#>  9 17915 PL            139   2295  2000 "Dispõe sobre a jornada de trabalho dos~
#> 10 18420 PL            139   2639  2000 "Dispõe sobre a atividade de curta dura~
#> # ... with 90 more rows